% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygkg.R
\name{querygkg}
\alias{querygkg}
\title{Performs a query on the Google Knowledge Graph API}
\usage{
querygkg(
  query = NULL,
  ids = NULL,
  lang = NULL,
  types = NULL,
  indent = NULL,
  prefix = NULL,
  limit = NULL,
  api.key
)
}
\arguments{
\item{query}{A character string to search for in Google Knowledge Graph.}

\item{ids}{A list of one or more entity IDs to search for in Google Knowledge Graph. Should ba a Google Knowledge Graph or Freebase ID code in character string format.}

\item{lang}{The list of language codes (defined in ISO 639) to run the query with, for instance `en`.}

\item{types}{Restricts returned entities to those of the specified types. See schema.org for valid types (e.g. `Person` as defined in http://schema.org/Person restricts the results to entities representing people). If multiple types are specified, returned entities will contain one or more of these types.}

\item{indent}{Enables indenting of JSON results.}

\item{prefix}{Enables prefix (initial substring) match against names and aliases of entities. For example, a prefix `Jung` will match entities and aliases such as `Jung`, `Jungle`, and `Jung-ho Kang`.}

\item{limit}{A numeric value limiting the number of entities to be returned. Maximum is 500. Default is 20. Requests with high limits have a higher chance of timing out.}

\item{api.key}{A Google API key.}
}
\value{
Returns the Google Knwoledge Graph output.
}
\description{
Performs a query on the Google Knowledge Graph API
}
\examples{
#Run a text-based query for the term "apple"
#querygkg(query = "apple", api.key = "YOUR_API_KEY")

#Run an ID-based query for the entity "apple"
#querygkg(ids = "/m/014j1m", api.key = "YOUR_API_KEY")
}
